💥 Adjust App's led variables
**Breaking change:**

Renamed `led*` constructor parameters and properties to be more
descriptive:
 - `ledOff` is now `ledHalt`, as it's only used when the badger is
   halted while on usb power.
 - `ledLow` is now `ledInactive` as it's used when the app is waiting
   for user input.
 - `ledHigh` is now `ledActive` as it's used when the app is processing
   the user input.

**Non breaking change:**

The default value for `ledActive` (formally `ledHigh`) changed from
`255` to `0`, this is to try and fix a bug causing artifacting to
appear on the screen when updating on battery power. I think it's being
caused by the board not having enough power when on battery to update
the screen correctly while also shining the LED at full power.